Flatten AbstractVsTextViewFilter and the generic derivation
There is now nothing in the generic version of AbstractVsTextViewFilter,
so we can flatten the two types in the hierarchy.
